Update #2 - testing bug of Thymeleaf 3 decoupled logic
It figured out that the issue mentioned previously in the commit 7a8a138 not to be a bug of Thymeleaf engine.
`@WebMvcTest` doesn't load customized auto-configuration, in this case which is `Thymeleaf3Config`.
I keep using normal Thymeleaf attributes in template, and keep decoupled logic code inside annotation block for later use.

* index.th.xml: added `#th-navbar-brand` attribute which was missing.