Add #10 - caching on shortenURL()
This uses Spring Cache to cache `shortenURL()` method of URLShorteningServiceImpl
to speed up getting the same short URL.
This actually skips the internal logic that checks if the input url already exists in the repository,
however this commit keep the code because the cache spec. like capacity or duration is not yet confirmed.